att python OBC

+++ oneBitCode/pythonFund/5.2-ex1.py
@@ -1,7 +1,7 @@
-num1 = int(input("Digite o primeiro número: "))
-num2 = int(input("Digite o segundo número: "))
-num3 = int(input("Digite o terceiro número: "))
-num4 = int(input("Digite o quarto número: "))
+num1 = float(input("Digite o primeiro número: "))
+num2 = float(input("Digite o segundo número: "))
+num3 = float(input("Digite o terceiro número: "))
+num4 = float(input("Digite o quarto número: "))
 
 media = (num1 + num2 + num3 + num4) / 4
